move head-data to addon
